+++ soulseek-rs-lib/src/message/handlers.rs
@@ -4,11 +4,11 @@ use crate::message::Message;
 use std::sync::mpsc::Sender;
 
 pub trait MessageHandler<Op>: Send {
-    fn get_code(&self) -> u8;
+    fn get_code(&self) -> u32;
     fn handle(&self, message: &mut Message, sender: Sender<Op>);
 }
 pub struct Handlers<Op> {
-    handlers: HashMap<u8, Box<dyn MessageHandler<Op> + Send>>,
+    handlers: HashMap<u32, Box<dyn MessageHandler<Op> + Send>>,
 }
 
 impl<Op> Default for Handlers<Op> {
@@ -35,7 +35,7 @@ impl<Op> Handlers<Op> {
     #[must_use]
     pub fn get_handler(
         &self,
-        code: u8,
+        code: u32,
     ) -> Option<&(dyn MessageHandler<Op> + Send)> {
         self.handlers.get(&code).map(|v| &**v)
     }
