+++ soulseek-rs-lib/src/message/peer/queue_upload.rs
@@ -8,7 +8,7 @@ use std::sync::mpsc::Sender;
 pub struct QueueUploadHandler;
 
 impl MessageHandler<PeerMessage> for QueueUploadHandler {
-    fn get_code(&self) -> u8 {
+    fn get_code(&self) -> u32 {
         43
     }
